Fix compilation warning with gcc 3.4.2 on Fedora Core 3 x86_64. This is effectively the same as the netplug-1.2.7-compiler.patch from Fedora Core 3.

+++ netlink.c
@@ -93,7 +93,7 @@ receive(int fd, struct msghdr *msg, int *status)
 
     if (msg->msg_namelen != sizeof(struct sockaddr_nl)) {
 	do_log(LOG_ERR, "Unexpected sender address length: got %d, expected %d",
-	       msg->msg_namelen, sizeof(struct sockaddr_nl));
+	       msg->msg_namelen, (int) sizeof(struct sockaddr_nl));
 	return done;
     }
